Updates for the provisioner and the ubuntu_install recipies.
Don't assume that pxelinux.0 is always located on the same place on the CD.
Instead, ensure it gets installed on the admin node, and copy it into place
from the location we know is installed if it is missing.

Don't assume that digest auth is enabled.  Only try to pass the machine install
key to the compute nodes if we are actually using digest auth.