Remove the DHCP -> Sledgehammer state machine.
Instead, drive that state via symlinks to the appropriate boot states
in the pxelinux.cfg dir.  Each MAC address we discover points to what
we think the next state should be, and we update symlinks instead of
rewriting DHCP configs and bouncing the DHCP server.  This reduces
the number of restarts to the DHCP server to once per host discovery
and delete.