Remove iPXE support.
iPXE is not as well-baked as it needs to be for our purposes -- it works
well enough in a VM test env, but it is failing in subtle and weird ways
on real hardware.

The iPXE stuff will be kept around on vlowther-ipxe for now for later
hacling once we have a chance to work out more of the issues that we were
seeing in the field.