control.sh: wait for pxe config to update before rebooting
It's possible the pxe config hasn't updated on the admin node
before the client reboots after it's allocated.  In this case
the client boots up in discovery again, and is "stuck" (well,
you can allocate again, and potentially repeat ad-infinitum...)

Most of the time this will never happen, but just in case,
this change uses curl to make sure the 01-$mac link exists
on the admin server before reboot.  If it doesn't exist, it
polls every 10 seconds for up to five minutes before giving
up and rebooting.