Correctly configure chef-client even if it was already installed
Instead of only configuring it when it's not installed, also check if
the installed config is the sample config. In that case, we can
overwrite it.

Also save the old config file, to be on the safe side.

(cherry picked from commit e3fb6a26acc24d0338e8fb95cb5a7e4d883fd6d9)