Move some of the PXE state machine from Chef to Crowbar.
This moves resposibility for updating the provisioner state for a node
from the update_nodes recipe into the provisioner service.

This lays some groundwork for hetero OS support by making it easier to
fail if we cannot install an OS.

It also lays some groundwork for automating UEFI installation by
directly exposing what PXE state the provisioner thinks they are in.