avoid endless installation loops
don't post installing state inside of autoyast, but set it from the
sleshammer immage. This will make sure to not going through more than
two installations. At the end of the installation, "installed" state is
posted. In case the admin server doesn't update the pxeconfig by then,
it will definitely be done once the second installation is done :(