Remove hardcoded dependency on usin /dev/sda as the install target.
Instead of hoping that /dev/sda will always be the device that the
system will boot to (which is just not true for several reasons, chief
among them that SCSI device enumeration is just not stable in any
system that has more than one disk controller), we now allow the
barclamps to pick a disk that the system should boot from and ensure
that the OS (including the bootloader) will be installed on this disk.

compute.ks.erb, net_seed.ks.erb, and autoyast.xml.erb have been
modified to let the provisioner feed a boot device into them.  By
default, this disk will be the one under /dev/disks/by-id
that corresponds to the first disk listed in /dev/disks/by-path. In
the normal course of things, this will ensure that we always get a
stable pointer to the first bootable device on the system no matter
what order modules are loaded in or what order PCI devices are
enumerated in.

See the bootdisk.rb recipe in the provisioner cookbook for more
information.

The bootdisk recipe has been designed so that it will not do anything
if node[:crowbar_wall][:boot_device] is already set.  This will allow
for future enhancements to the RAID and BIOS barclamps to pick the
boot device of their choosing.

It is strongly advised that node[:crowbar_wall][:boot_device] be a
device in the /dev/disk/by-id directory, as these names are based on
the unique ATA, SCSI, SATA, or SAS identifiers for the drive, and not
based on the vagaries of SCSI numbering or device naming.