Allow SSh access to discovered nodes.
The provisioner barclamp already allows to specify additional ssh access
keys for the admin node and it makes sense to just reuse those for nodes
that are in discovery state. Helps with debugging and is much easier
than the current solution based on hooks:

  $ cat /updates/discovered-pre/passwd.hook
  PASS="1r21lb1dp1ssw4rd"
  echo -e "$PASS\n$PASS" | passwd

Also, ~/.ssh/authorized_keys mode should be 0644, not 0700.