Fix boot controller finding logic for the C8220
In order to boot properly when we want to use both the on-board
AHCI controller (for the OS) and an LSO9205 (for external SATA
storage).  We need to do this because one of the expansion slots is
hanging off a lower PCI address than the baked-in AHCI controller on
the C8220.