Added support for deploying windows clients from the provisioner
Added a parameter (target_platform) that will indicate if windows (and which flavor)
should be installed.
Ensured that any services that can not be deployed on windows nodes (e.g. ipmi,
bluepill, ssh) and any features that are not available (e.g. core_dump) will be skipped.
Added the tftpd.conf file required for handling uppercase-lowercase naming issues on
Windows (the installer for windows does many requests with capital first letter or capital
letters)
Updated the tftpd bluepill template to make sure that the tftpd.conf file is loaded as required.
Added the template for the unattended.xml (that ensures that the OS gets installed without
user intervention).