parse_node_data: allow key elements containing '.', by backslash-escaping
    ./parse_node_data -a 'foo.2\.0.bar' test.json

will now return 'value' when fed:

    {
      "foo": {
        "2.0": {
          "bar": "value"
        }
      }
    }

This is now required since we will be enumerating repositories to be
included in autoyast.xml generation from meta-data contained in
provisioner.json, and we want to key that meta-data by repository
name to avoid requiring an extra set of repository keys.  Repository
names can be things like "SUSE-Cloud-2.0-Updates", hence the need for
allowing JSON key elements to contain ".".

There is a corresponding commit in the top-level crowbar repository
adding the same behaviour to json-edit.