Add openstack-keystone Should-Start in SUSE init script for crowbar_join
crowbar_join will call chef-client; if this is happening on a controller
node, then the server roles for glance/cinder/quantum/nova will try to
contact the keystone API. But this one won't have started yet since the
keystone cookbook will only restart it delayed (unless the config file
changed, which won't happen on boot).

So we need to have keystone running before.