Try harder in skipping the MSFT compat rules
The by-id/ MS compatibility hack links that were added
as a maintenance update (and thereby are available in
the discovery image) are not recognized by the SP3 install
media, hence missing during install, which breaks.

The previous regexp was not good enough, some harddrives
do not start their naming with a letter, but a number. In
this case it seems udev always prefixes it with "2", so
hopefully this regexp is good enough..

https://bugzilla.novell.com/show_bug.cgi?id=864857