Avoid infinite loops with sntp
We do it 120 times, and stop looping afterwards.

It is quite a long time in the worst case: timeout seems to be 15
seconds, so that'd be ~30 minutes. However, worst case should be avoided
since we don't even enter the loop if the NTP port is not open on the
server.

If the NTP server replies but without time, that means we try during 2
minutes, which is not too bad.

(cherry picked from commit 57991054344b31b8b35e66191ff3415d272c0cce)