Update dhcp host files that we don't manage anymore
If, for some reason, a node has less MAC addresses than before, we need
to drop the DHCP host files that we won't update with chef anymore.

As the host files have a format like
"d52-54-00-21-c3-72.site-$index.conf", we simply remove the host files
where the index is too high.