crowbar_join.suse: Reset reboot flag after reboot
Currently we reset node[:crowbar_wall][:wait_for_reboot] always before a
chef-client run. That doesn't work if there's another chef-client run
before the reboot happens.
Add a check which compares the stored time when the reboot was requested
and the time the node was booted to verify that the reboot finished.