Don't use by-path links for virtio disks
When using virtio disk and /dev/disk/by-id/ links don't exist for it (no serial
number configured for the disk) do not fall back to the /dev/disk/by-path
links.  These are unstable and while they are present in the discovery image,
they might not exist anymore in the deployed OS (e.g. when deploying SLES12,
which uses a more recent udev version).