Set target platform before moving to hardware-installing (bnc#913867)
... only when it's not set already, obviously.

We can't do that in the cookbook on the server side, so let's do it in
the rails app.

The reason for not doing it in allocate is that it doesn't work when
crowbar_register is used. And we need it in hardware-installing because
we will claim disks for boot.

https://bugzilla.suse.com/show_bug.cgi?id=913867