Make SUSE Cloud repos optional to allow Storage-only deployments
This makes the SUSE Cloud repos optional, in the same way that SLE HA
and Storage repos are already optional, i.e. everything except the base
OS is now optional.

If the Storage repos are present but Cloud repos are not, the autoyast
profile is tweaked to perform a Storage-only deployment, by installing
ses-release instead of suse-openstack-cloud-release.

Signed-off-by: Tim Serong <tserong@suse.com>