bsc#980569 update option: -r is now -S
https://bugzilla.suse.com/show_bug.cgi?id=980569

from the old manpage:
-r
steps the system clock to the correct time by the Unix settimeofday
system call. Requires root priviledge.
http://doc.ntp.org/4.2.6p5/sntp.html

from the new manpage:
-S, --step
By default, sntp displays the clock offset but does not attempt to
correct it. This option enables offset correction by stepping, that is,
directly setting the clock to the corrected time. This typically
requires sntp be invoked as the superuser ("root"). Note that the short
option is -S, an uppercase letter S.
https://www.eecis.udel.edu/~mills/ntp/html/sntp.html