Use ntpdate on systems with newer ntpd (bsc#980569)
With a recent update to ntpd-4.2.8p6 the commandline options of sntp changed in
a backwards incompatible way. Try to find out with version of ntp we're running
and uses sntp for older releases (4.2.4 didn't contain a working ntpdate) and
ntpdate for newer releases to do the initial time sync.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=980569